1.支持了一些Attribute。可以定义NodeWidth和NodeColor啦！ 2.定义NodeColor后，选中节点的样式消失了；Attribute还剩下一个没有实现RequireNodes的功能

+++ Assets/Scripts/SugarNodeUIToolkit/NodeDataBase/Attribute/NodeAttributes/NodeWidth.cs
@@ -5,12 +5,12 @@ namespace SugarNode
     [AttributeUsage(AttributeTargets.Class, Inherited = true, AllowMultiple = false)]
     public sealed class NodeWidthAttribute : Attribute
     {
-        public uint width;
+        public float width;
         /// <summary>
         /// 在网格空间下指定节点的宽度
         /// </summary>
         /// <param name="width">1代表宽为1个网格（粗稀线）</param>
-        public NodeWidthAttribute(uint width)
+        public NodeWidthAttribute(float width)
         {
             this.width = width;
         }
